Use symbolic icon in  tray
Icons are being removed from Plasma themes and imported into breeze.
To avoid having the panel icon appear as the app icon it was renamed
to kruler-symbolic in breeze icons. Because of the icon theme lookup
rules, icon themes without a "kruler-symbolic" will use a "kruler" icon.
When running on an older Plasma with a "kruler" override this will
also be used instead of "kruler-symbolic".